pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer"
authors = ["Denis <snegirev.dv@gmail.com>"]
readme = "README.md"
packages = [
    { include = "page_analyzer" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "License :: Freeware"
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
validators = "^0.28.3"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
